agentic-systems/pyproject.toml
[project]
name = "agentic-systems"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "langchain>=0.3.28",
    "langchain-openai>=0.3.35",
    "langchain-snowflake>=0.2.2",
    "langgraph>=0.6.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "snowflake-connector-python>=4.3.0",
]
